requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
scikit-learn>=0.24.2
matplotlib>=3.4.3

# Data fetching
yfinance>=0.1.63
requests>=2.26.0
beautifulsoup4>=4.9.0
lxml>=4.9.0

# Web interface
streamlit>=0.84.0
plotly>=5.3.1
transformers>=4.11.0

# Configuration and environment
python-dotenv>=0.19.0
joblib>=1.0.1
tqdm>=4.62.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.0
pytest-mock>=3.6.0

# Development tools
black>=21.7b0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0

# Documentation
sphinx>=4.1.0
sphinx-rtd-theme>=0.5.0

# Added from the code block
seaborn>=0.11.2